chore: Moved CompositionObject ContextStore to a new file.
